@@ -1,10 +1,12 @@
 {
   "name": "pocketbase-superuser-ui",
   "private": true,
+  "packageManager": "pnpm@9.11.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json || true"
   },
   "type": "module",
   "prettier": {
@@ -13,28 +15,34 @@
     "svelteBracketNewLine": true
   },
   "devDependencies": {
-    "@codemirror/autocomplete": "^6.0.0",
-    "@codemirror/commands": "^6.0.0",
-    "@codemirror/lang-html": "^6.1.0",
-    "@codemirror/lang-javascript": "^6.0.2",
-    "@codemirror/lang-json": "^6.0.0",
-    "@codemirror/lang-sql": "^6.4.0",
-    "@codemirror/language": "^6.0.0",
-    "@codemirror/legacy-modes": "^6.0.0",
-    "@codemirror/search": "^6.0.0",
-    "@codemirror/state": "^6.0.0",
-    "@codemirror/view": "^6.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
-    "chart.js": "^4.0.0",
-    "chartjs-adapter-luxon": "^1.2.0",
+    "@codemirror/autocomplete": "^6.18.0",
+    "@codemirror/commands": "^6.6.0",
+    "@codemirror/lang-html": "^6.4.8",
+    "@codemirror/lang-javascript": "^6.2.2",
+    "@codemirror/lang-json": "^6.0.1",
+    "@codemirror/lang-sql": "^6.7.1",
+    "@codemirror/language": "^6.10.8",
+    "@codemirror/legacy-modes": "^6.4.1",
+    "@codemirror/search": "^6.5.8",
+    "@codemirror/state": "^6.4.1",
+    "@codemirror/view": "^6.34.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@tailwindcss/forms": "^0.5.9",
+    "autoprefixer": "^10.4.20",
+    "chart.js": "^4.4.3",
+    "chartjs-adapter-luxon": "^1.3.1",
     "chartjs-plugin-zoom": "^2.0.1",
+    "daisyui": "^4.12.14",
     "leaflet": "^1.9.4",
-    "luxon": "^3.4.4",
+    "luxon": "^3.5.0",
     "pocketbase": "^0.25.1",
-    "sass": "^1.45.0",
-    "svelte": "^4.0.0",
+    "postcss": "^8.4.47",
+    "sass": "^1.77.8",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.3",
     "svelte-flatpickr": "^3.3.3",
-    "svelte-spa-router": "^4.0.0",
-    "vite": "^5.0.11"
+    "svelte-spa-router": "^4.0.2",
+    "tailwindcss": "^3.4.10",
+    "vite": "^5.4.3"
   }
 }
